Primeira Finalização

+++ src/petshop/modelos/Servico.java
@@ -21,7 +21,7 @@ public class Servico {
         return preco;
     }
 
-    public void setPreco(int preco) {
+    public void setPreco(double preco) {
         this.preco = preco;
     }
